requirements.txt
azure-functions
requests
beautifulsoup4
google-generativeai
python-dotenv
azure-cosmos
